feat/telegram

+++ app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, TIMESTAMP, func
+from sqlalchemy import Column, Integer, String, TIMESTAMP, func, BIGINT 
 from sqlalchemy.orm import relationship
 from database import Base 
 
@@ -6,7 +6,7 @@ class User(Base):
     __tablename__ = 'users'
 
     userId = Column(Integer, primary_key=True, autoincrement=True)
-    telegramId = Column(Integer, unique=True, nullable=False)
+    telegramId = Column(BIGINT, unique=True, nullable=False)
     accessToken = Column(String, nullable=True)
     refreshToken = Column(String, nullable=True)
     createdAt = Column(TIMESTAMP, server_default=func.now())
